Add check for local changes in obs-api package using rpm
Add a playbook that will fail the deployment if there are local changes
to the package obs-api. The verification is done via rpm and in my tests
the changes are printed if the shell command has non zero return code.

Requires a bool (overwrite_local_changes, default to false) to switch
mode in case the local changes need to be overwritten.

Since I cannot fully test, for this draft I'm adding the include just in
deploy_without_migration.yml for completion.